Warning: big commits (but it is the easier) 1) merge back macro specific localization files => generic localization files (idem for templates) 2) During this merge, I strip the double quote ("" => ", '' => ') to a single quote. It is working like this in the native code. 3) Get ride of the macros specific localization files 4) Update the "make localization" target to not use locales_macros/ 5) Update the updateLocalizationModule.sh script to add more checks I decided to skip the temporary generation in the locales_macros directory to perform the stripping only on the macros. It would have broke Java or C/C++ code containing '' for example
The script produced to perform this task:
------------
MODULES=`find . -maxdepth 1 -type d  ! -name ".*" ! -name 'javasci'`
LANG="de_DE.po  fr_FR.po  ja_JP.po  pt_BR.po  uk_UA.po  zh_TW.po ca_ES.po          es_ES.po  it_IT.po  pl_PL.po  ru_RU.po  zh_CN.po"

for m in $MODULES; do
    m=$(echo $m|sed -e "s|\./||g")
    cd $m
    echo "in $(pwd)"
    for l in $LANG; do
        if test -f locales_macros/$l; then
            sed -i -e "s/\"\"/\\\"/g" -e "s/''/'/g" -e "s/'\"/\\\"/g" locales_macros/$l
            sed -i -e "s/msgstr \"$/msgstr \"\"/" -e "s/msgid \"$/msgid \"\"/" locales_macros/$l

            if test ! -f locales/$l; then
                # no native locale
                cp locales_macros/$l locales/$l
            else
                msgcat --use-first -o locales/$l.tmpsly locales_macros/$l locales/$l
                if test $? -ne 0; then
                    exit 32
                fi
                mv locales/$l.tmpsly locales/$l
            fi
        fi
    done
    cd -
done
--------

+++ scilab/tools/localization/updateLocalizationModule.sh
@@ -116,10 +116,10 @@ function process_module {
     if test $IS_MACROS -eq 1; then
         generate_find_command EXTENSIONS_MACROS
         local TARGETDIR=$TARGETDIR_MACROS
-        if test ! -d $TARGETDIR; then mkdir $TARGETDIR; fi
     else
         generate_find_command EXTENSIONS
     fi
+    if test ! -d $TARGETDIR; then mkdir $TARGETDIR; fi
 
     FILES=`eval $FILESCMD|tr "\n" " "|sort`
 
@@ -159,17 +159,20 @@ function process_module {
     fi
 
     $XGETTEXT $XGETTEXT_OPTIONS -p $TARGETDIR/ -o $MODULE_NAME.pot.tmp $FILES > /dev/null
-
-    # We need C strings format to be used as gettext key
-    # "" -> \"
-    # '' -> '
-    # '" -> \"
-    # "' -> '
-    sed -e "s/\"\"/\\\"/g" -e "s/''/'/g" -e "s/'\"/\\\"/g" -e "s/\"'/'/g" $TARGETDIR/$MODULE_NAME.pot.tmp > $TARGETDIR/$MODULE_NAME.pot.tmp2
-    # We introduced invalid tag [msgstr "] and [msgid "]
-    # restore them [msgstr ""] and [msgid ""]
-    sed -e "s/msgstr \"$/msgstr \"\"/" -e "s/msgid \"$/msgid \"\"/" $TARGETDIR/$MODULE_NAME.pot.tmp2 > $TARGETDIR/$MODULE_NAME.pot.tmp
-    rm $TARGETDIR/$MODULE_NAME.pot.tmp2 2> /dev/null
+    if test ! -f $MODULE_NAME.pot.tmp -a $IS_MACROS -eq 1; then
+        # Empty file => no string found
+        # We are modifing on the fly Scilab localization files
+        # 
+        # We need C strings format to be used as gettext key
+        # "" -> \"
+        # '' -> '
+        # '" -> \"
+        # "' -> ' -e "s/\"'/'/g" 
+        sed -i -e "s/\"\"/\\\"/g" -e "s/''/'/g" -e "s/'\"/\\\"/g" $TARGETDIR/$MODULE_NAME.pot.tmp
+        # We introduced invalid tag [msgstr "] and [msgid "]
+        # restore them [msgstr ""] and [msgid ""]
+        sed -i -e "s/msgstr \"$/msgstr \"\"/" -e "s/msgid \"$/msgid \"\"/" $TARGETDIR/$MODULE_NAME.pot.tmp
+    fi
 
     if test  -z "$CreationDate"; then
         # File not existing before ... Set the current date a POT-Creation-Date
@@ -177,11 +180,32 @@ function process_module {
     else
         sed -e "s/MODULE/$MODULE_NAME/" -e "s/CREATION-DATE/$CreationDate/" -e "s/REVISION-DATE/`date +'%Y-%m-%d %H:%M'`$TIMEZONE/" $HEADER_TEMPLATE > $LOCALIZATION_FILE_US
     fi
-    cat $LOCALIZATION_FILE_US.tmp >> $LOCALIZATION_FILE_US
+
+    echo "LOCALIZATION_FILE_US : $LOCALIZATION_FILE_US"
+    msguniq -u $LOCALIZATION_FILE_US.tmp >> $LOCALIZATION_FILE_US
     rm $LOCALIZATION_FILE_US.tmp 2> /dev/null
+
+    MSGOUTPUT=$(msgcat $LOCALIZATION_FILE_US)
+    if test $? -ne 0; then
+        echo "Badly formated localization files"
+        exit 32
+    fi
     if test -z "$(msgcat $LOCALIZATION_FILE_US)"; then
         # empty template. Kill it!
         rm $LOCALIZATION_FILE_US
+    fi
+
+    if test $IS_MACROS -eq 1; then
+        LOCALIZATION_FILE_NATIVE=$(echo $LOCALIZATION_FILE_US|sed -e "s|-macros||g" -e "s|_macros||g")
+        if test ! -f $LOCALIZATION_FILE_NATIVE; then
+            # no native code. Copy the macro one
+            cp $LOCALIZATION_FILE_US $LOCALIZATION_FILE_NATIVE
+        else
+        # merge locale macros => native code
+            msgcat --use-first  -o $LOCALIZATION_FILE_NATIVE.tmp  $LOCALIZATION_FILE_NATIVE $LOCALIZATION_FILE_US
+            mv $LOCALIZATION_FILE_NATIVE.tmp $LOCALIZATION_FILE_NATIVE
+        fi
+        rm -rf $TARGETDIR_MACROS
     fi
 
     # Remove fake file used to extract string from XML
@@ -202,6 +226,6 @@ for MODULE in $MODULES; do
     cd $PATHTOPROCESS
     process_module "src"
     process_module "macros"
-
+    
     cd $SCI/
 done # Browse modules
